refactor(runner): move graceful kill process to lib/utils
